Added Managers and project structure refactor
